Reference

Here, you can see the documentation for each type and method provided by AuditoryNerveFiber.jl. Higher-level functions designed for end users are documented first, while lower-level functions designed mostly for internal are documented later.

Wrappers

Wrappers are functions that "wrap" around lower-level bindings and/or implementations and provide more convenience than the bindings themselves. For example, when calling wrappers a user should never have to worry about handling pointers, pre-allocating arrays, etc. These are the functions that most users will want to use in their own code. Presently, the wrappers provide access to the Zilany, Bruce, and Carney (2014) auditory-nerve model and all share the same basic signature:

function_name(input, cf; kwargs...)

where the input is a 1D array, cf is a characteristic frequency in Hz, and kwargs are various other parameters that have default values (i.e., you can set them if you want to, but if you don't a default value will be used). Below, you can see each of the wrappers documented below, but you can also head over to Examples to see examples of how these functions can be used.

AuditoryNerveFiber.sim_ihc_zbc2014 — Function

sim_ihc_zbc2014(input, cf; fs=10e4, cohc=1.0, cihc=1.0, species="human", n_rep=1)

Simulates inner hair cell potential for given acoustic input.

Arguments

input::Vector{Float64}: sound pressure waveform in pascals

cf::Float64: characteristic frequency of the IHC in Hz

fs::Float64: sampling rate in Hz

cohc::Float64: outer hair cell survival (from 0 to 1)

cihc::Float64: inner hair cell survival (from 0 to 1)

species::String: species, either ("cat" = cat, "human" = humans with Shera tuning, "human_glasberg" = humans with Glasberg tuning)

n_rep::Int64: how many repetitions to perform. Because the inner hair cell stage has no randomness, n_rep > 1 simply concatenates n_rep copies of IHC simulation and returns

Returns

output::Vector{Float64}: inner hair cell potential output, size of length(input)*n_rep

AuditoryNerveFiber.sim_synapse_zbc2014 — Function

sim_synapse_zbc2014(input, cf; fs=10e4, fs_synapse=10e3, fiber_type="high", power_law="approximate", fractional=false, n_rep=1)

Simulates synapse output for a given inner hair cell input

Arguments

input::Vector{Float64}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fs::Float64: sampling rate of the input in Hz

fs_synapse::Float64: sampling rate of the interior synapse simulation. The ratio between fs and fs_synapse must be an integer.

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

power_law::String: whether we use true or approximate power law adaptation, one of ("actual", "approximate")

fractional::Bool: whether we use ffGn or not, one of (true, false)

n_rep::Int64: number of repetititons to run. We assume that the input was also generated using n_rep=n_rep, hence we infer that the input acoustic waveform is of length length(input)/n_rep.

Returns

output::Vector{Float64}: synapse output (unknown units?), length is length(input)

AuditoryNerveFiber.sim_an_zbc2014 — Function

sim_an_zbc2014(input, cf; fs=10e4, fs_synapse=10e3, power_law="approximate", fractional=false, n_rep=1)

Simulates auditory nerve output (spikes and firing rate) for a given inner hair cell input

Arguments

input::Vector{Float64}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fs::Float64: sampling rate of the input in Hz

fs_synapse::Float64: sampling rate of the interior synapse simulation. The ratio between fs and fs_synapse must be an integer.

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

power_law::String: whether we use true or approximate power law adaptation, one of ("actual", "approximate")

fractional::Bool: whether we use ffGn or not, one of (true, talse)

n_rep::Int64: number of repetititons to run. We assume that the input was also generated using n_rep=n_rep, hence we infer that the input acoustic waveform is of length length(input)/n_rep.

Returns

meanrate::Vector{Float64}: analytical estimate of instantaneous firing rate, of size length(input)/n_rep

varrrate::Vector{Float64}: analytical estimate of instantaneous firing rate variance, of size length(input)/n_rep

psth::Vector{Float64}: peri-stimulus time histogram, of size length(input)/n_rep

Notes

The behavior of n_rep in sim_an_zbc2014 and derived functions differs slightly from

sim_synapse_zbc2014. sim_synapse_zbc2014 will return an array of length length(input), and will contain a single response to n_rep concatenated copies of the inner hair cell potential to a single copy of the underlying acoustic stimulus (assuming that sim_ihc_zbc2014 was also evaluated using n_rep=n_rep). sim_an_zbc2014 will return an array of length length(input). Underneath, it generates the same synapse response as sim_synapse_zbc2014, but then averages over the repetitions and averages the spike train into a peristimulus time histogram (PSTH).

AuditoryNerveFiber.sim_anrate_zbc2014 — Function

sim_anrate_zbc2014(input, cf; fs=10e4, fs_synapse=10e3, power_law="approximate", fractional=false)

Simulates auditory nerve output (instantaneous firing rate) for a given inner hair cell input

Arguments

input::Vector{Float64}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fs::Float64: sampling rate of the input in Hz

fs_synapse::Float64: sampling rate of the interior synapse simulation. The ratio between fs and fs_synapse must be an integer.

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

power_law::String: whether we use true or approximate power law adaptation, one of ("actual", "approximate")

fractional::Bool: whether we use ffGn or not, one of (true, talse)

n_rep::Int64: number of repetititons to run

Returns

psth::Vector{Float64}: analytical estimate of instantaneous firing rate, size of length(input) (see docs for sim_an_zbc2014 to understand why)

AuditoryNerveFiber.sim_spikes_zbc2014 — Function

sim_spikes_zbc2014(input, cf; fs=10e4, fs_synapse=10e3, power_law="approximate", fractional=false, n_rep=1)

Simulates auditory nerve output (spikes only) for a given inner hair cell input

Arguments

input::Vector{Float64}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fs::Float64: sampling rate of the input in Hz

fs_synapse::Float64: sampling rate of the interior synapse simulation. The ratio between fs and fs_synapse must be an integer.

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

power_law::String: whether we use true or approximate power law adaptation, one of ("actual", "approximate")

fractional::Bool: whether we use ffGn or not, one of (true, talse)

n_rep::Int64: number of repetititons to run

Returns

psth::Vector{Float64}: peri-stimulus time histogram, size of length(input)/n_rep (see docs for sim_an_zbc2014 to understand why)

Bindings

Bindings are functions that directly interface with external implementations of models in other languages. All bindings here work generally the same way:

Inputs are passed directly to external functions with minimal or no input checking or preprocessing

No defaults are provided

Names and types of variables match those of the original-language source code

Generally, it is expected that end-users will never need to call these functions. Instead, they should use a wrapper or another higher-level function to access the same functionality in a a more "user-friendly" way.

AuditoryNerveFiber.BM! — Function

BM!(px, cf, nrep, tdres, totalstim, cohc, cihc, species, ihcout, bmout)

Direct binding to BM C function in model_IHC.c

Passes arguments directly to BM using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

px::Vector{Float64}: sound pressure waveform in pascals

cf::Float64: characteristic frequency of the fiber in Hz

nrep::Int32: number of repetitions to simulate. Note that for the IHC simulation, one "true" simulation is conducted and then that simulation is copied and tiled (because there is no randomness in the IHC simulation) to simulate multiple times.

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

totalstim::Int32: number of samples in simulation

cohc::Float64: outer hair cell survival (from 0 to 1)

cihc::Float64: inner hair cell survival (from 0 to 1)

species::Int32: species, either (1 = cat, 2 = humans with Shera tuning, 3 = humans with Glasberg tuning)

ihcout::Vector{Float64}: array of same size as px, used to store IHC output from C

bmout::Vector{Float64}: array of same size as px, used to store BM output from C

AuditoryNerveFiber.IHCAN! — Function

IHCAN!(px, cf, nrep, tdres, totalstim, cohc, cihc, species, ihcout)

Direct binding to IHCAN C function in model_IHC.c

Passes arguments directly to IHCAN using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

px::Vector{Float64}: sound pressure waveform in pascals

cf::Float64: characteristic frequency of the fiber in Hz

nrep::Int32: number of repetitions to simulate. Note that for the IHC simulation, one "true" simulation is conducted and then that simulation is copied and tiled (because there is no randomness in the IHC simulation) to simulate multiple times.

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

totalstim::Int32: number of samples in simulation

cohc::Float64: outer hair cell survival (from 0 to 1)

cihc::Float64: inner hair cell survival (from 0 to 1)

species::Int32: species, either (1 = cat, 2 = humans with Shera tuning, 3 = humans with Glasberg tuning)

ihcout::Vector{Float64}: array of same size as px, used to store output from C

AuditoryNerveFiber.Synapse! — Function

Synapse!(ihcout, tdres, cf, totalstim, nrep, spont, noiseType, implnt, sampFreq, synouttmp)

Direct binding to Synapse C function in model_Synapse.c

Passes arguments directly to Synapse using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

ihcout::Vector{Float64}: output from IHC simulation (IHCAN!)

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

cf::Float64: characteristic frequency of the fiber in Hz

totalstim::Int32: number of samples in simulation

nrep::Int32: number of repetitions to simulate.

spont::Float64: spontaneous rate, either (0.1 == low spont fiber, 4.0 == medium spont fiber, 100.0 == high spont fiber)

noiseType::Float64: whether we use ffGn or not (1.0 == ffGn, 0.0 == not)

implnt::Float64: whether or not to use exact implementation of power-law adaptation, either (1.0 == actual, 0.0 == approximate)

sampFreq::Float64: sampling frequency of the power law stage in Hz. Simulations are decimated to sampFreq from 1/tdres before the power law stage and then upsampled back to the original sampling rate. The product of tdres and sampFreq, which indicates the amount to decimate by, must be an integer

synouttmp::Vector{Float64}: array of same size as ihcout, used to store output from C

AuditoryNerveFiber.SingleAN! — Function

SingleAN!(ihcout, cf, nrep, tdres, totalstim, fibertype, noiseType, implnt, meanrate, varrate, psth)

Direct binding to Synapse C function in model_Synapse.c

Passes arguments directly to Synapse using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

ihcout::Vector{Float64}: output from IHC simulation (IHCAN!)

cf::Float64: characteristic frequency of the fiber in Hz

nrep::Int32: number of repetitions to simulate.

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

totalstim::Int32: number of samples in simulation

fibertype::Float64: fiber type, either (1.0 == low, 2.0 == med, 3.0 == high)

noiseType::Float64: whether we use ffGn or not (1.0 == ffGn, 0.0 == not)

implnt::Float64: whether or not to use exact implementation of power-law adaptation, either (1.0 == actual, 0.0 == approximate)

meanrate::Vector{Float64}: array of same size as ihcout, used to store analytical firing rate output

varrate::Vector{Float64}: array of same size as ihcout, used to store analytical firing rate variance output

psth::Vector{Float64}: array of same size as ihcout, used to store empirical PSTH output