Reference

Here, you can see the documentation for each type and method provided by AuditoryNerveFiber.jl. Higher-level functions designed for end users are documented first, while lower-level functions designed mostly for internal are documented later.

Wrappers

Wrappers are functions that "wrap" around lower-level bindings and/or implementations and provide more convenience than the bindings themselves. For example, when calling wrappers a user should never have to worry about handling pointers, pre-allocating arrays, etc. These are the functions that most users will want to use in their own code. Presently, the wrappers provide access to the Zilany, Bruce, and Carney (2014) auditory-nerve model and all share the same basic signature:

func(input, cf; kwargs)

where the input is a 1D array, cf is a characteristic frequency in Hz, and kwargs are various other parameters that have default values. Below, you can see each of the wrappers documented below, but you can also head over to Examples to see examples of how these functions can be used.

AuditoryNerveFiber.sim_ihc_zbc2014 — Function

sim_ihc_zbc2014(input, cf; fs=10e4, cohc=1.0, cihc=1.0, species="cat")

Simulates inner hair cell potential for given acoustic input.

Arguments

input::Array{Float64, 1}: sound pressure waveform in pascals

cf::Float64: characteristic frequency of the IHC in Hz

fs::Float64: sampling rate in Hz

cohc::Float64: outer hair cell survival (from 0 to 1)

cihc::Float64: inner hair cell survival (from 0 to 1)

species::Int32: species, either (1 = cat, 2 = humans with Shera tuning, 3 = humans with

Glasberg tuning)

Returns

output::Array{Float64, 1}: inner hair cell potential output

AuditoryNerveFiber.sim_synapse_zbc2014 — Function

sim_synapse_zbc2014(input, cf; fs=10e4, fs_synapse=10e3, cohc=1.0, cihc=1.0)

Simulates synapse output for a given inner hair cell input

Arguments

input::Array{Float64, 1}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fs::Float64: sampling rate of the input in Hz

fs_synapse::Float64: sampling rate of the interior synapse simulation. The ratio between

fs and fs_synapse must be an integer.

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

frac_noise::String: controls whether we use true or approximate fractional Gaussian

noise implementation, one of ("actual", "approximate")

Returns

output::Array{Float64, 1}: synapse output (unknown units?)

AuditoryNerveFiber.sim_an_zbc2014 — Function

sim_an_zbc2014(input, cf; fs=10e4, fiber_type="high", cohc=1.0, cihc=1.0)

Simulates auditory nerve output (spikes or firing rate) for a given inner hair cell input

Arguments

input::Array{Float64, 1}: input hair cell potential (from simihczbc2014)

cf::Float64: characteristic frequency of the fiber in Hz

fiber_type::String: fiber type, one of ("low", "medium", "high") spontaneous rate

frac_noise::String: controls whether we use true or approximate fractional Gaussian

noise implementation, one of ("actual", "approximate")

Returns

meanrate::Array{Float64, 1}: analytical estimate of instantaneous firing rate

varrrate::Array{Float64, 1}: analytical estimate of instantaneous firing rate variance

psth::Array{Float64, 1}: peri-stimulus time histogram (NOT IMPLEMENTED, SHOULD BE EMPTY)

Bindings

Bindings are functions that directly interface with external implementations of models in other languages. All bindings here work generally the same way:

Inputs are passed directly to external functions with minimal or no input checking or preprocessing

No defaults are provided

Names and types of variables match those of the original-language source code

Generally, it is expected that end-users will never need to call these functions. Instead, they should use a wrapper or another higher-level function to access the same functionality in a a more "user-friendly" way.

AuditoryNerveFiber.IHCAN! — Function

IHCAN!(px, cf, nrep, tdres, totalstim, cohc, cihc, species, ihcout)

Direct binding to IHCAN C function in model_IHC.c

Passes arguments directly to IHCAN using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

px::Array{Float64, 1}: sound pressure waveform in pascals

cf::Float64: characteristic frequency of the fiber in Hz

nrep::Int32: number of repetitions to simulate. Note that for the IHC simulation,

one "true" simulation is conducted and then that simulation is copied and tiled (because there is no randomness in the IHC simulation) to simulate multiple times.

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

totalstim::Int32: number of samples in simulation

cohc::Float64: outer hair cell survival (from 0 to 1)

cihc::Float64: inner hair cell survival (from 0 to 1)

species::Int32: species, either (1 = cat, 2 = humans with Shera tuning, 3 = humans

with Glasberg tuning)

ihcout::Array{Float64, 1}: array of same size as px, used to store output from C

AuditoryNerveFiber.Synapse! — Function

Synapse!(ihcout, tdres, cf, totalstim, nrep, spont, noiseType, implnt, sampFreq, 
         synouttmp)

Direct binding to Synapse C function in model_Synapse.c

Passes arguments directly to Synapse using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

ihcout::Array{Float64, 1}: output from IHC simulation (IHCAN!)

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

cf::Float64: characteristic frequency of the fiber in Hz

totalstim::Int32: number of samples in simulation

nrep::Int32: number of repetitions to simulate.

spont::Float64: spontaneous rate, either (0.1 == low spont fiber, 4.0 == medium spont

fiber, 100.0 == high spont fiber)

noiseType::Float64: NOT CURRENTLY IMPLEMENTED

implnt::Float64: whether or not to use exact implementation of fractional Gaussian

noise, either (1.0 == use, 0.0 == approximate)

sampFreq::Float64: sampling frequency of the power law stage in Hz. Simulations are

decimated to sampFreq from 1/tdres before the power law stage and then upsampled back to the original sampling rate. The product of tdres and sampFreq, which indicates the amount to decimate by, must be an integer

synouttmp::Array{Float64, 1}: array of same size as ihcout, used to store output from

C

AuditoryNerveFiber.SingleAN! — Function

SingleAN!(ihcout, cf, nrep, tdres, totalstim, fibertype, noiseType, implnt, meanrate, 
          varrate, psth)

Direct binding to Synapse C function in model_Synapse.c

Passes arguments directly to Synapse using ccall. Arrays are converted to pointers, functions are converted to pointers, and all other types are converted directly to corresponding types in C. Note that while there are type checks enforced automatically by Julia, there are no sanity checks on any arguments.

Arguments

ihcout::Array{Float64, 1}: output from IHC simulation (IHCAN!)

cf::Float64: characteristic frequency of the fiber in Hz

nrep::Int32: number of repetitions to simulate.

tdres::Float64: time-domain resolution (i.e., reciprocal of sampling rate)

totalstim::Int32: number of samples in simulation

fibertype::Float64: fiber type, either (1.0 == low, 2.0 == med, 3.0 == high)

noiseType::Float64: NOT CURRENTLY IMPLEMENTED

implnt::Float64: whether or not to use exact implementation of fractional Gaussian

noise, either (1.0 == use, 0.0 == approximate)

meanrate::Array{Float64, 1}: array of same size as ihcout, used to store analytical

firing rate output

varrate::Array{Float64, 1}: array of same size as ihcout, used to store analytical

firing rate variance output

psth::Array{Float64, 1}: array of same size as ihcout, used to store empirical PSTH

output